Dev/Python/pumpControlCenter/requirements.txt
pyFirmata==1.1.0
pyserial==3.5
tinydb==4.7.0
